Separate tests into multiple files
