+++ __tests__/classes/Bindings.spec.ts
@@ -1,0 +1,8 @@
+import { FeeRate } from '../../src/classes/Bindings';
+
+describe('FeeRate', () => {
+  const feeRate = new FeeRate(10);
+  it('Should return a double when called', () => {
+    expect(feeRate.asSatPerVb()).toBe(10);
+  });
+});
