+++ __tests__/setup.ts
@@ -1,0 +1,1 @@
+jest.mock('./../src/index.ts');
